+++ CHANGELOG.md
@@ -13,6 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `not`, `tag`
 - `fail` creates a labelled failure, which is like a throw for parsers
 - `failLabel(parser, label)` is a shortcut for `or(parser, fail(label))`
-- `cstToAst` transforms a CST to an AST
+- `cstToIst` transforms a CST to an IST
 
 [unreleased]: https://github.com/unleashy/vahv/compare/v0.1.0...HEAD